requirements.txt
django==5.0
django-filter==23.5
djangorestframework==3.14.0
markdown==3.5.1
requests
pdfminer_six